+++ examples2d/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod character;
+pub mod svg;