+++ src/geometry/collider.rs
@@ -1,4 +1,6 @@
 use crate::dynamics::{CoefficientCombineRule, MassProperties, RigidBodyHandle};
+#[cfg(feature = "dim3")]
+use crate::geometry::HeightFieldFlags;
 use crate::geometry::{
     ActiveCollisionTypes, BroadPhaseProxyIndex, ColliderBroadPhaseData, ColliderChanges,
     ColliderFlags, ColliderMassProps, ColliderMaterial, ColliderParent, ColliderPosition,
@@ -10,10 +12,8 @@ use crate::pipeline::{ActiveEvents, ActiveHooks};
 use crate::prelude::ColliderEnabled;
 use na::Unit;
 use parry::bounding_volume::{Aabb, BoundingVolume};
-use parry::shape::{Shape, TriMeshBuilderError, TriMeshFlags};
-
-#[cfg(feature = "dim3")]
-use crate::geometry::HeightFieldFlags;
+use parry::shape::{Shape, TriMeshBuilderError, TriMeshFlags, VoxelPrimitiveGeometry};
+use parry::transformation::voxelization::FillMode;
 
 #[cfg_attr(feature = "serde-serialize", derive(Serialize, Deserialize))]
 #[derive(Clone, Debug)]
@@ -571,6 +571,55 @@ impl ColliderBuilder {
     /// of its planar boundary.
     pub fn halfspace(outward_normal: Unit<Vector<Real>>) -> Self {
         Self::new(SharedShape::halfspace(outward_normal))
+    }
+
+    /// Initializes a shape made of voxels.
+    ///
+    /// Each voxel has the size `voxel_size` and grid coordinate given by `centers`.
+    /// The `primitive_geometry` controls the behavior of collision detection at voxels boundaries.
+    ///
+    /// For initializing a voxels shape from points in space, see [`Self::voxels_from_points`].
+    /// For initializing a voxels shape from a mesh to voxelize, see [`Self::voxelized_mesh`].
+    pub fn voxels(
+        primitive_geometry: VoxelPrimitiveGeometry,
+        voxel_size: Vector<Real>,
+        voxels: &[Point<i32>],
+    ) -> Self {
+        Self::new(SharedShape::voxels(primitive_geometry, voxel_size, voxels))
+    }
+
+    /// Initializes a collider made of voxels.
+    ///
+    /// Each voxel has the size `voxel_size` and contains at least one point from `centers`.
+    /// The `primitive_geometry` controls the behavior of collision detection at voxels boundaries.
+    pub fn voxels_from_points(
+        primitive_geometry: VoxelPrimitiveGeometry,
+        voxel_size: Vector<Real>,
+        points: &[Point<Real>],
+    ) -> Self {
+        Self::new(SharedShape::voxels_from_points(
+            primitive_geometry,
+            voxel_size,
+            points,
+        ))
+    }
+
+    /// Initializes a voxels obtained from the decomposition of the given trimesh (in 3D)
+    /// or polyline (in 2D) into voxelized convex parts.
+    pub fn voxelized_mesh(
+        primitive_geometry: VoxelPrimitiveGeometry,
+        vertices: &[Point<Real>],
+        indices: &[[u32; DIM]],
+        voxel_size: Real,
+        fill_mode: FillMode,
+    ) -> Self {
+        Self::new(SharedShape::voxelized_mesh(
+            primitive_geometry,
+            vertices,
+            indices,
+            voxel_size,
+            fill_mode,
+        ))
     }
 
     /// Initialize a new collider builder with a cylindrical shape defined by its half-height